1part_4: Checking whether a number belongs to a range
The program prompts the user to enter a number from 10 to 20. If the entered value is less than 10, the message "The number is less than the required range" is displayed and prompts you to enter the number again. If the entered value is greater than 20, the message "The number is greater than the required range" is displayed and asks to try again. The program works until a value from the range from 10 to 20 is entered, after which the message "Thank you" will be displayed.