1part_5: Determining the verbal result of the game
The program solves the following problem:
At the football championship, a team is given 3 points for a win, 0 for a loss, 1 for a draw. The number of points received by the team for the game is known. The program outputs the verbal result of the game (win, loss or draw). The number is entered from the console.