2part_1: Color matching, which was chosen by the user t computer
This program outputs the names of five colors, randomly the program selects one and offers to do the same to the user. If the user chooses the same color that the program has chosen, the message "Excellent!" is displayed; otherwise, the answer is displayed, which hides a hint of the correct color. If the colors do not match, the program also prompts the user to try again; if the user does not guess this time, the same hint is displayed again and asked to choose a color (and so on until the user gives the correct answer).