2part_2: Adding another program to a certain position to the list of TV programs
This program creates a list with the names of four TV shows and displays them on separate lines. The program prompts the user to enter the name of another transmission and the position at which it should be inserted into the list. The program displays a list in which all five gears are in new positions.