2part_4: Solving the quadratic equation
The values of the coefficients a, b, c are entered from the keyboard. The program checks the value of the discriminant: if it is less than zero, there are no roots, if it is zero, then the root is 1, if it is greater than zero, there are two roots. To do this, a construction of the form if elif else is used.