2part_6: Optimizing car travel
The program asks the user for the following data: How many kilometers do you want to drive by car? How many liters of fuel does a car consume per 100 kilometers? How many liters of fuel is in your tank? Then, depending on the entered values, the program gives a verdict: whether the user will travel the desired distance or not.